+++ lib/src/carvings/string/append.dart
@@ -1,11 +1,13 @@
-import '../../interface.dart';
+import 'package:carvable/src/carvings/positional.dart';
 
-class CarvingAppend extends Carving<String, String> {
+class CarvingAppend extends PositionalCarving<String, String> {
 	final String replacement;
-  final int? at;
-	CarvingAppend(this.replacement, { this.at });
+  final bool tailing;
+	CarvingAppend(this.replacement, { int? at }) :
+    tailing = at == null,
+    super(at ?? 0);
 	
-	@override String apply(String input) => at == null
+	@override String apply(String input, { int offset = 0 }) => tailing
     ? input + replacement
-    : input.replaceRange(at!, at!, replacement);
+    : input.replaceRange(start + offset, start + offset, replacement);
 }